+++ tests/data/pgsql.sql
@@ -123,21 +123,21 @@ CREATE TABLE "order" (
   created_at integer NOT NULL,
   updated_at integer NOT NULL,
   deleted_at integer,
-  total decimal(10,0) NOT NULL
+  total float NOT NULL
 );
 
 CREATE TABLE "order_with_null_fk" (
   id serial not null primary key,
   customer_id integer,
   created_at integer NOT NULL,
-  total decimal(10,0) NOT NULL
+  total float NOT NULL
 );
 
 CREATE TABLE "order_item" (
   order_id integer NOT NULL references "order"(id) on UPDATE CASCADE on DELETE CASCADE,
   item_id integer NOT NULL references "item"(id) on UPDATE CASCADE on DELETE CASCADE,
   quantity integer NOT NULL,
-  subtotal decimal(10,0) NOT NULL,
+  subtotal float NOT NULL,
   PRIMARY KEY (order_id, item_id)
 );
 
@@ -145,7 +145,7 @@ CREATE TABLE "order_item_name" (
   order_id integer NOT NULL,
   item_name varchar(128) NOT NULL,
   quantity integer NOT NULL,
-  subtotal decimal(10,0) NOT NULL,
+  subtotal float NOT NULL,
   PRIMARY KEY (order_id, item_name)
 );
 
@@ -153,7 +153,7 @@ CREATE TABLE "order_item_with_null_fk" (
   order_id integer,
   item_id integer,
   quantity integer NOT NULL,
-  subtotal decimal(10,0) NOT NULL
+  subtotal float NOT NULL
 );
 
 CREATE TABLE "composite_fk" (
